Assignment2
Created Assignment2 Scene which has 2 modes. One without AR and one with AR. In both modes you can spin and scale the 2 objects. There are sliders to do those actions. There is a button to switch objects between the donut (which I made) and the die (which I didnt make). There is also a button to switch between AR and Non AR mode